RAB-1284: rollback bad changes
