+++ src/Akeneo/Pim/Structure/.php_cd.php
@@ -18,7 +18,6 @@ $rules = [
         'Akeneo\Platform\Bundle\FeatureFlagBundle\FeatureFlags',
         'Oro\Bundle\SecurityBundle\Annotation\AclAncestor',
         'Oro\Bundle\SecurityBundle\SecurityFacade',
-        'Oro\Bundle\SecurityBundle\Exception\AccessDeniedException',
         'FOS\RestBundle\View',
         'Oro\Bundle\DataGridBundle\Event\BuildBefore',
         'Oro\Bundle\FilterBundle\Grid\Extension\Configuration',
