+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/handler.yml
@@ -2,4 +2,4 @@ services:
 
   Akeneo\Pim\Structure\Bundle\Application\GetAttributeGroup\GetAttributeGroupHandler:
     arguments:
-      - '@Akeneo\Pim\Structure\Bundle\infrastructure\Query\Sql\GetAttributeGroups'
+      - '@Akeneo\Pim\Structure\Bundle\Infrastructure\Query\Sql\GetAttributeGroups'
