+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/queries.yml
@@ -102,6 +102,10 @@ services:
         arguments:
             - '@database_connection'
 
+    Akeneo\Pim\Structure\Bundle\Query\InternalApi\AttributeGroup\Sql\GetAttributeGroups:
+        arguments:
+            - '@database_connection'
+
     akeneo.pim.structure.query.get_association_type_translations:
         class: Akeneo\Pim\Structure\Bundle\Query\PublicApi\Association\Sql\SqlGetAssociationTypeTranslations
         arguments:
@@ -200,11 +204,6 @@ services:
     Akeneo\Pim\Structure\Bundle\Query\InternalApi\Family\GetFamilyIdsNotUsedByProductsQuery:
         arguments:
             - '@database_connection'
-
-    Akeneo\Pim\Structure\Bundle\infrastructure\Query\Sql\GetAttributeGroups:
-        arguments:
-            - '@database_connection'
-            - '@feature_flags'
 
     Akeneo\Pim\Structure\Bundle\Query\InternalApi\AttributeOption\FindAttributeOptions:
         class: Akeneo\Pim\Structure\Bundle\Query\InternalApi\AttributeOption\SqlFindAttributeOptions
