+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/models/AttributeGroup.ts
@@ -7,7 +7,6 @@ type AttributeGroup = {
   sort_order: number;
   labels: AttributeGroupLabels;
   is_dqi_activated: boolean;
-  selected: boolean;
 };
 
 type AttributeGroupCollection = {
